[Feat] 전체 상품 조회 페이징 추가

+++ src/main/java/com/example/ecommerce/dto/product/ProductResponse.java
@@ -6,7 +6,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 public record ProductResponse(
     @Schema(description = "상품 ID") Long id,
     @Schema(description = "상품명") String name,
-    @Schema(description = "가격") Integer price,
+    @Schema(description = "가격") Long price,
     @Schema(description = "상세 설명") String description,
     @Schema(description = "재고 수량") Integer stockQty,
     @Schema(description = "상품 상태") ProductStatus status
